Fix bugs

+++ db/migrations/2016_07_09_025132_create_hourly_quotes_table.py
@@ -9,7 +9,7 @@ class CreateHourlyQuotesTable(Migration):
         """
         with self.schema.create('hourly_quotes') as table:
             table.increments('id')
-            table.double('price')
+            table.float('price')
             table.date('date')
             table.integer('stock_id').unsigned()
             table.foreign('stock_id').references('id').on('stocks')
